Changes Windows version to directly use `CreateFileA`
`CreateFile` is `#define`d as `CreateFileW` in unicode builds (when `/D
UNICODE` is defined). This is probably more likely to be a issue when
integrating into GUI projects.

This commit allows compilation to succeed when `UNICODE` has been
defined and adds a note to the README, explicitly stating that mio
doesn't support wide char string types as parameters.